refactor: move About.getDetailString string function into separate module
